requirements.txt
fastapi==0.110.0
pydantic==2.9.2
pydantic-settings==2.5.2
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
uvicorn==0.29.0
pymongo==4.8.0
python-dotenv==1.0.1
joblib==1.3.2
numpy==1.26.4
pandas==2.2.0
psutil==6.0.0
scikit-learn==1.4.2
tensorflow==2.17.0